tests: Update unit tests
